update script fil in case t1 or t2 is missing

+++ prepare_seg_and_gt.sh
@@ -41,24 +41,26 @@ cd $PATH_DATA_PROCESSED/data/$SUBJECT/anat/
 contrast='T1 T2'
 for i in $contrast; do
 	file=${SUBJECT}_${i}w
-	c_args=${i/T/t}
+	if test -f $PATH_DATA/$SUBJECT/anat/${file}.nii.gz;then
+		c_args=${i/T/t}
 
-	## copy needed file (t1w and T2w other are not needed) 
-	cp $PATH_DATA/$SUBJECT/anat/${file}.nii.gz ./
-	## Deepseg to get needed segmentation. 
-	sct_deepseg_sc -i ${file}.nii.gz -c ${c_args}  -ofolder $PATH_DATA_PROCESSED/data/derivatives/labels/$SUBJECT/anat/
+		## copy needed file (t1w and T2w other are not needed) 
+		cp $PATH_DATA/$SUBJECT/anat/${file}.nii.gz ./
+		## Deepseg to get needed segmentation. 
+		sct_deepseg_sc -i ${file}.nii.gz -c ${c_args}  -ofolder $PATH_DATA_PROCESSED/data/derivatives/labels/$SUBJECT/anat/
 
-	## seg file name
-	file_seg=$PATH_DATA_PROCESSED/data/derivatives/labels/$SUBJECT/anat/${SUBJECT}_${i}w_seg.nii.gz
-	label_file=$PATH_DATA_PROCESSED/data/derivatives/labels/$SUBJECT/anat/${SUBJECT}_${i}w_labels-disc-manual.nii.gz
+		## seg file name
+		file_seg=$PATH_DATA_PROCESSED/data/derivatives/labels/$SUBJECT/anat/${SUBJECT}_${i}w_seg.nii.gz
+		label_file=$PATH_DATA_PROCESSED/data/derivatives/labels/$SUBJECT/anat/${SUBJECT}_${i}w_labels-disc-manual.nii.gz
 
-	sct_label_vertebrae -i ${file}.nii.gz -s ${file_seg} -c t2 -discfile ${label_file} -ofolder $PATH_DATA_PROCESSED/data/derivatives/labels/$SUBJECT/anat/
+		sct_label_vertebrae -i ${file}.nii.gz -s ${file_seg} -c t2 -discfile ${label_file} -ofolder $PATH_DATA_PROCESSED/data/derivatives/labels/$SUBJECT/anat/
 
-	## Change the name to avoid overwriting files output by sct_label_vertebrae during prediction later. 
-	mv $PATH_DATA_PROCESSED/data/derivatives/labels/$SUBJECT/anat/${SUBJECT}_${i}w_seg_labeled_discs.nii.gz $PATH_DATA_PROCESSED/data/derivatives/labels/$SUBJECT/anat/${SUBJECT}_${i}w_projected-gt.nii.gz
+		## Change the name to avoid overwriting files output by sct_label_vertebrae during prediction later. 
+		mv $PATH_DATA_PROCESSED/data/derivatives/labels/$SUBJECT/anat/${SUBJECT}_${i}w_seg_labeled_discs.nii.gz $PATH_DATA_PROCESSED/data/derivatives/labels/$SUBJECT/anat/${SUBJECT}_${i}w_projected-gt.nii.gz
 
-	## Chage data type to int 16 for "projected_gt"
-	sct_image -i $PATH_DATA_PROCESSED/data/derivatives/labels/$SUBJECT/anat/${SUBJECT}_${i}w_projected-gt.nii.gz -type int16
-
+		## Chage data type to int 16 for "projected_gt"
+		sct_image -i $PATH_DATA_PROCESSED/data/derivatives/labels/$SUBJECT/anat/${SUBJECT}_${i}w_projected-gt.nii.gz -type int16
+	
+	fi
 done
